Major UI modernization and Notepad++ feature set
- Upgrade to WPF UI (Fluent Design) with FluentWindow and Mica backdrop
- Add 5 new color themes: Dracula, Monokai, OneDark, SolarizedDark, SolarizedLight
- Refine existing themes with syntax highlighting color support
- Implement Find & Replace with custom SearchPanel template (Ctrl+F/H)
- Add Go To Line dialog (Ctrl+G)
- Add Word Wrap and Show Whitespace toggles
- Add Zoom In/Out/Reset (Ctrl++/-/0 and Ctrl+Scroll)
- Add drag-and-drop file open
- Add Recent Files MRU (10 entries)
- Add Save As (Ctrl+Shift+S)
- Add line operations: Duplicate (Ctrl+D), Move Up/Down (Alt+Up/Down)
- Add Undo/Redo to Edit menu
- Fix tab close button with code-behind Click handler
- Fix slow app close by skipping async cycle for clean tabs
- Fix frozen brush crash in search panel marker
- Fix core library robustness (LruCache thread safety, hex editor)

+++ src/FastEdit.Core/FileAnalysis/BinaryDetector.cs
@@ -75,7 +75,7 @@ public class BinaryDetector
                 bufferSize: SampleSize,
                 useAsync: true);
 
-            var buffer = new byte[Math.Min(SampleSize, stream.Length)];
+            var buffer = new byte[(int)Math.Min(SampleSize, stream.Length)];
             var bytesRead = await stream.ReadAsync(buffer.AsMemory());
 
             if (bytesRead == 0)
